+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 2.2.0
+
+### Minor Changes
+
+- [#1169](https://github.com/synergy-design-system/synergy-design-system/pull/1169) [`aefb7a8`](https://github.com/synergy-design-system/synergy-design-system/commit/aefb7a8c4c9860f6222d7054a6f044b2ed0c49a6) Thanks [@schilchSICKAG](https://github.com/schilchSICKAG)! - Released on: 2026-02-02
+
+  chore: 🔧 remove vendorism from components package (#1168)
+
+  The custom vendoring process is not needed anymore. Because of this, it was removed and all comments and builds now only rely on `pnpm build`
+
+## 2.1.0
+
+### Minor Changes
+
+- [#1141](https://github.com/synergy-design-system/synergy-design-system/pull/1141) [`96dff2e`](https://github.com/synergy-design-system/synergy-design-system/commit/96dff2ebb47cee901f72773664ee864db5653219) Thanks [@schilchSICKAG](https://github.com/schilchSICKAG)! - Released on: 2026-02-02
+
+  feat: ✨ `<syn-button-group>` (#392)
+
+  We added a new component `<syn-button-group>` that you can use to group a list of buttons.
+
+  It will automatically apply its `size` and `variant` properties to all nested buttons.
+
 ## 2.0.0
 
 ### Major Changes